requirements.txt
click==8.1.7
Jinja2==3.1.4
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdit-plain==1.0.1
mdurl==0.1.2
python-frontmatter==1.1.0
PyYAML==6.0.1
